+++ src/gis_jp.js
@@ -1,4 +1,4 @@
-import { elevationJapan as usage } from './modules/usage'
+import { gis_jp as usage } from './modules/usage'
 import { elevation } from './modules/elevation'
 import { ignoredBlocks } from './modules/blocks'
 import { transformIDs } from './modules/utils'